requirements.txt
httplib2~=0.18.1
oauth2client~=4.1.3
google-api-python-client~=1.10.0
google-cloud-bigquery~=1.26.0
pandas~=1.0.5
flask~=1.1.2
gspread-dataframe~=3.0.7
gspread-formatting~=0.2.5
gspread~=3.6.0
gunicorn~=20.0.4
requests~=2.24.0
beautifulsoup4~=4.9.1
nltk~=3.5
rake-nltk~=1.0.4
praw~=7.1.0
